replacing my types.h with default types

+++ src/color.c
@@ -1,9 +1,9 @@
 #include <limits.h>
 #include <stdarg.h>
 #include <stdio.h>
+#include <stdlib.h>
 
 #include "color.h"
-#include "types.h"
 
 #define BGDIFF 10
 #define GETBIT(B, P) (((B) >> (P)) & 1)
@@ -20,8 +20,8 @@ enum effects {
   _CROSSOUT, // \x1b[9m
 };
 
-static i32 apply_color(FILE *stream, struct style *s) {
-  i32 written_bytes = 0;
+static int apply_color(FILE *stream, struct style *s) {
+  int written_bytes = 0;
 
   written_bytes += fprintf(stream ? stream : stdout, 
       ESC SEPARATOR "%u" TERMINATOR, s->foreground);
@@ -31,11 +31,11 @@ static i32 apply_color(FILE *stream, struct style *s) {
   return written_bytes;
 }
 
-static i32 apply_effects(FILE *stream, struct style *s) {
-  i32 written_bytes = 0;
+static int apply_effects(FILE *stream, struct style *s) {
+  int written_bytes = 0;
 
-  u64 totalbits = sizeof(s->effects) * CHAR_BIT;
-  for (u64 i = 0; i < totalbits; i++) {
+  size_t totalbits = sizeof(s->effects) * CHAR_BIT;
+  for (size_t i = 0; i < totalbits; i++) {
     if (GETBIT(s->effects, i)) {
       switch (i) {
         case _BOLD:
@@ -77,11 +77,11 @@ static i32 apply_effects(FILE *stream, struct style *s) {
   return written_bytes;
 }
 
-i32 printfc(struct style s, const char *fmt, ...) {
+int printfc(struct style s, const char *fmt, ...) {
   va_list args;
   va_start(args, fmt);
 
-  i32 written_bytes = 0;
+  int written_bytes = 0;
   written_bytes += apply_color(NULL, &s);
   written_bytes += apply_effects(NULL, &s);
   written_bytes += vprintf(fmt, args);
@@ -91,11 +91,11 @@ i32 printfc(struct style s, const char *fmt, ...) {
   return written_bytes;
 }
 
-i32 fprintfc(FILE *stream, struct style s, const char *fmt, ...) {
+int fprintfc(FILE *stream, struct style s, const char *fmt, ...) {
   va_list args;
   va_start(args, fmt);
 
-  i32 written_bytes = 0;
+  int written_bytes = 0;
   written_bytes += apply_color(stream, &s);
   written_bytes += apply_effects(stream, &s);
   written_bytes += vfprintf(stream, fmt, args);
